+++ src/Symfony/Bundle/SecurityBundle/CHANGELOG.md
@@ -23,6 +23,7 @@ CHANGELOG
  * Add `token_endpoint_auth_method` option (`client_secret_post`, `client_secret_basic`) to the `oidc_login` authenticator
  * Add `enable_end_session` and `post_logout_redirect_path` options to the `oidc_login` authenticator for RP-Initiated Logout
  * Add `direct_redirect` option to the `oidc_login` authenticator
+ * Add the `none` value to the `token_endpoint_auth_method` option of the `oidc_login` authenticator, to declare a public client, and make `client_secret` optional for it
 
 8.1
 ---